requirements.txt
Flask==3.1.0
gunicorn==23.0.0
mysql-connector-python==9.1.0
textblob==0.19.0
matplotlib==3.10.0
